backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.26.0
pgvector==0.2.4
supabase==2.3.4
huggingface-hub==0.20.2
transformers==4.36.2
python-dotenv==1.0.0

